Update Time!!
- Now checks before downloading CCleaner installer
- Glary is included with the Malwarebytes ninite installer
- Now has the option to make the classic NOC rescue folder with icon and everything
- Supports downloading the rescue.msi for the NOC Calling card, hosted in this github account. Can be used for initial instillation or just repairing

+++ AutoTune/Repository/PSAScript/SetupTools.ps1
@@ -1,27 +1,42 @@
-#Set-ExecutionPolicy RemoteSigned -Scope CurrentUser
-
-#mkdir %installpath%\TempDownloads
-
 # Download Installers for Tools
-
-echo "Downloading CCleaner. . ."
-curl -O "C:\Users\Public\Downloads\CCSetup.exe" "https://bits.avcdn.net/productfamily_CCLEANER/insttype_FREE/platform_WIN_PIR/installertype_ONLINE/build_RELEASE"
-echo "Download Complete!"
-echo ""
-
-echo "Downloading Glary Utilities. . ."
-curl -o "C:\Users\Public\Downloads\GlarySetup.exe" "https://www.glarysoft.com/aff/download.php?s=GU"
-echo "Download Complete!"
-echo ""
 
 echo "Downloading ADW Cleaner. . ."
 curl -o "C:\Users\Public\Downloads\ADWCleaner.exe" "https://adwcleaner.malwarebytes.com/adwcleaner?channel=release"
 echo "Download Complete!"
 echo ""
 
-echo "Downloading Malwarebytes. . ."
-curl -o "C:\Users\Public\Downloads\MBSetup.exe" "https://ninite.com/malwarebytes/ninite.exe"
-echo "Download Complete!"
-echo ""
+If (-NOT (Test-Path "C:\Users\Public\Downloads\CCSetup.exe" -PathType Leaf)) {
+    echo "Downloading CCleaner. . ."
+    curl -O "C:\Users\Public\Downloads\CCSetup.exe" "https://bits.avcdn.net/productfamily_CCLEANER/insttype_FREE/platform_WIN_PIR/installertype_ONLINE/build_RELEASE"
+    echo "Download Complete!"
+    echo ""
+
+}
+else {
+    echo "CCleaner Already Downloaded!"
+}
+
+If (-NOT (Test-Path "C:\Users\Public\Downloads\Glary&MBSetup.exe" -PathType Leaf)) {
+    echo "Downloading Glary & Malwarebytes. . ."
+    curl -o "C:\Users\Public\Downloads\Glary&MBSetup.exe" "https://ninite.com/glary-malwarebytes/ninite.exe"
+    echo "Download Complete!"
+    echo ""
+}
+else {
+    echo "Glary and Malwarebytes Already Downloaded!"
+}
+
+If (-NOT (Test-Path "C:\Users\Public\Downloads\remote.msi" -PathType Leaf)) {
+    echo "Downloading NOC Remote Support Installer. . ."
+    curl -o "C:\Users\Public\Downloads\InstallerMain.zip" "https://github.com/AGiggleSniffer/Installers-for-Autotune/archive/refs/heads/main.zip"
+    Expand-Archive -Path "C:\Users\Public\Downloads\InstallerMain.zip" -DestinationPath "C:\Users\Public\Downloads"
+    Move-Item –Path "C:\Users\Public\Downloads\Installers-for-Autotune-main\Noc_Downloads\remote.msi" -Destination "C:\Users\Public\Downloads"
+    Remove-Item -Path "C:\Users\Public\Downloads\Installers-for-Autotune-main","C:\Users\Public\Downloads\InstallerMain.zip" -Recurse
+    echo "Download Complete!"
+    echo ""
+}
+else {
+    echo "Remote.msi Already Downloaded!"
+}
 
 #Read-Host -Prompt "Press any key to continue"
